+++ p2panda-rs/src/schema/error.rs
@@ -6,8 +6,20 @@ use thiserror::Error;
 #[derive(Error, Debug)]
 pub enum SchemaIdError {
     /// Invalid hash in schema id.
-    #[error("invalid hash string")]
+    #[error("encountered invalid hash while parsing application schema id: {0}")]
     HashError(#[from] crate::hash::HashError),
+
+    /// Encountered a malformed schema id.
+    #[error("malformed schema id: {0}")]
+    MalformedSchemaId(String),
+
+    /// Application schema ids must start with the schema's name.
+    #[error("application schema id is missing a name: {0}")]
+    MissingApplicationSchemaName(String),
+
+    /// Invalid system schema id.
+    #[error("not a known system schema: {0}")]
+    UnknownSystemSchema(String),
 }
 
 /// Custom errors related to `Schema`.