+++ CHANGELOG.md
@@ -39,7 +39,10 @@ Highlights are marked with a pancake 🥞
 - Improve error reporting when adding operation fields [#262](https://github.com/p2panda/p2panda/issues/262)] `rs` `js`
 - Update mock node API [#286](https://github.com/p2panda/p2panda/issues/286) `rs`
 - Refactored graph module to be generic over graph node keys and other graph improvements [#289](https://github.com/p2panda/p2panda/issues/289) `rs`
-- Require sorted serialisation of document view ids [#284](https://github.com/p2panda/p2panda/pull/284) `js`
+- Require sorted serialisation of document view ids [#284](https://github.com/p2panda/p2panda/pull/284) `rs`
+- Introduce new application schema id format [#292](https://github.com/p2panda/p2panda/pull/292) `rs`
+- Update spelling of system schema ids [#294](https://github.com/p2panda/p2panda/pull/294) `rs`
+- Update `Schema` implementation to make use of new `SchemaId` [#296](https://github.com/p2panda/p2panda/pull/296) `rs`
 
 ## Fixed
 
@@ -53,6 +56,7 @@ Highlights are marked with a pancake 🥞
 - Split up overly long `operation.rs` file [#232](https://github.com/p2panda/p2panda/pull/232) `rs`
 - Extend test coverage for `OperationFields` [#236](https://github.com/p2panda/p2panda/pull/236) `rs`
 - Further develop our best practices for writing documentation [#240](https://github.com/p2panda/p2panda/pull/240) `rs`
+- Test `debug` macro calls in Github CI [#288](https://github.com/p2panda/p2panda/pull/288) `rs`
 
 ## [0.3.0]